# kin-kernel-20slot/service/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Kin Gateway v2 API
  version: 0.1.0
  description: Reference data-plane and control-plane contract.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/config/proxy-pool:
    servers: [{url: http://localhost:9090}]
    get:
      operationId: getProxyPoolConfig
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Typed proxy pool config. GET never includes credentials.
          content: {application/json: {schema: {$ref: "#/components/schemas/ProxyPoolEnvelope"}}}
    put:
      operationId: putProxyPoolConfig
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: "#/components/schemas/ProxyPoolPut"}}}
      responses:
        "200": {description: Proxy pool config committed.}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /api/v1/proxies:
    servers: [{url: http://localhost:9090}]
    get:
      operationId: listProxies
      security: [{InternalBearer: []}]
      responses:
        "200":
          description: Desensitized proxy list plus pool config and totals.
          content: {application/json: {schema: {$ref: "#/components/schemas/ProxyPoolSnapshot"}}}
  /api/v1/proxies/{id}/reveal:
    servers: [{url: http://localhost:9090}]
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    post:
      operationId: revealProxy
      security: [{InternalBearer: []}]
      parameters: [{$ref: "#/components/parameters/OperatorHeader"}]
      responses:
        "200":
          description: The only plaintext credential response. URI only. Do not cache.
          headers:
            Cache-Control: {schema: {type: string, enum: [no-store]}}
          content:
            application/json:
              schema:
                type: object
                required: [id, uri]
                additionalProperties: false
                properties:
                  id: {type: string}
                  uri: {type: string}
        "503": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    InternalBearer:
      type: http
      scheme: bearer
      description: KIN_CONTROL_INTERNAL_TOKEN shared only with trusted internal callers.
  parameters:
...
    OperatorHeader:
      name: x-kin-operator
      in: header
      required: false
      schema: {type: string, maxLength: 128}
      description: Trusted operator identity supplied by the Node BFF; defaults to control-api.
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [type, error]
      properties:
        type: {const: error}
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            retryable: {type: boolean}
            request_id: {type: string}
...
    ApplyStatus:
      type: object
      additionalProperties: false
      required: [state, desired_revision, effective_revision]
      properties:
        state: {type: string, enum: [effective, pending_restart, rejected]}
        desired_revision: {type: integer, minimum: 0}
        effective_revision: {type: integer, minimum: 0}
        last_error: {type: string, maxLength: 4096}
...
    ProxyPoolConfig:
      type: object
      additionalProperties: false
      required: [probe_interval_min, probe_timeout_ms, max_failures, enabled, disconnect_on_error, bind_limit]
      properties:
        probe_interval_min: {type: integer, enum: [5, 10, 30, 60]}
        probe_timeout_ms: {type: integer, minimum: 1000, maximum: 60000}
        max_failures: {type: integer, minimum: 1, maximum: 10}
        enabled: {type: boolean}
        disconnect_on_error: {type: boolean}
        bind_limit: {type: integer, minimum: 1, maximum: 32}
    ProxyPoolPut:
      type: object
      additionalProperties: false
      required: [expected_revision, data]
      properties:
        expected_revision: {type: integer, minimum: 0}
        data: {type: object, description: Deep partial ProxyPoolConfig patch.}
        import: {type: boolean, default: false}
    ProxyPoolEnvelope:
      type: object
      required: [schema_version, revision, config_hash, runtime_hash, kernel_hash, data, apply, degraded]
      properties:
        schema_version: {type: integer, minimum: 1}
        revision: {type: integer, minimum: 0}
        config_hash: {type: string, pattern: "^[a-f0-9]{64}$"}
        runtime_hash: {type: string, pattern: "^[a-f0-9]{64}$"}
        kernel_hash: {type: string, pattern: "^[a-f0-9]{64}$"}
        data: {$ref: "#/components/schemas/ProxyPoolConfig"}
        apply: {$ref: "#/components/schemas/ApplyStatus"}
        degraded: {type: boolean}
    PublicProxy:
      type: object
      additionalProperties: false
      required: [id, host, port, has_auth, status, enabled, bound_vm_ids]
      properties:
        id: {type: string}
        revision: {type: integer, minimum: 0}
        host: {type: string}
        port: {type: integer, minimum: 1, maximum: 65535}
        has_auth: {type: boolean}
        status: {type: string, enum: [unknown, ok, fail, dead]}
        enabled: {type: boolean}
        bound_vm_id: {type: [string, "null"]}
        bound_vm_ids: {type: array, items: {type: string}}
        bound_count: {type: integer}
        bind_limit: {type: integer}
        consecutive_failures: {type: integer}
        latency_ms: {type: [integer, "null"]}
        last_probe_at: {type: string}
        last_error: {type: string}
        created_at: {type: string, format: date-time}
    ProxyPoolSnapshot:
      type: object
      required: [config, totals, proxies]
      properties:
        config: {$ref: "#/components/schemas/ProxyPoolEnvelope"}
        totals: {type: object, additionalProperties: {type: integer}}
        proxies: {type: array, items: {$ref: "#/components/schemas/PublicProxy"}}
...